# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Job Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Resume {
  id: number;
  name: string;
  email?: string;
  skills: string[];
  experience_years?: number;
  summary?: string;
  analyzed_at?: string;
}

export interface ResumeInput {
  name: string;
  email?: string;
  raw_text: string;
  filename?: string;
}

export type PreferencesWorkType = typeof PreferencesWorkType[keyof typeof PreferencesWorkType];


export const PreferencesWorkType = {
  remote: 'remote',
  hybrid: 'hybrid',
  onsite: 'onsite',
  any: 'any',
} as const;

export interface Preferences {
  id: number;
  work_type: PreferencesWorkType;
  /** @nullable */
  city?: string | null;
  roles: string[];
  /** @nullable */
  min_salary?: number | null;
  updated_at?: string;
}

export type PreferencesInputWorkType = typeof PreferencesInputWorkType[keyof typeof PreferencesInputWorkType];


export const PreferencesInputWorkType = {
  remote: 'remote',
  hybrid: 'hybrid',
  onsite: 'onsite',
  any: 'any',
} as const;

export interface PreferencesInput {
  work_type: PreferencesInputWorkType;
  /** @nullable */
  city?: string | null;
  roles: string[];
  /** @nullable */
  min_salary?: number | null;
}

export type JobWorkType = typeof JobWorkType[keyof typeof JobWorkType];


export const JobWorkType = {
  remote: 'remote',
  hybrid: 'hybrid',
  onsite: 'onsite',
} as const;

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  pending: 'pending',
  applied: 'applied',
  ignored: 'ignored',
} as const;

export interface Job {
  id: number;
  title: string;
  company: string;
  location: string;
  work_type: JobWorkType;
  /** @nullable */
  salary?: number | null;
  /** @nullable */
  salary_text?: string | null;
  match_score: number;
  skills_match?: string[];
  skills_missing?: string[];
  /** @nullable */
  description?: string | null;
  /** @nullable */
  url?: string | null;
  status: JobStatus;
  /** @nullable */
  cover_letter?: string | null;
  found_at?: string;
}

export interface SearchResult {
  jobs_found: number;
  message: string;
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  applied: 'applied',
  interview: 'interview',
  offer: 'offer',
  rejected: 'rejected',
} as const;

export interface Application {
  id: number;
  job_id: number;
  job_title: string;
  company: string;
  status: ApplicationStatus;
  applied_at: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  salary?: number | null;
  work_type?: string;
  match_score?: number;
}

export interface ApplicationInput {
  job_id: number;
  /** @nullable */
  notes?: string | null;
}

export type ApplicationUpdateStatus = typeof ApplicationUpdateStatus[keyof typeof ApplicationUpdateStatus];


export const ApplicationUpdateStatus = {
  applied: 'applied',
  interview: 'interview',
  offer: 'offer',
  rejected: 'rejected',
} as const;

export interface ApplicationUpdate {
  status?: ApplicationUpdateStatus;
  /** @nullable */
  notes?: string | null;
}

export interface AgentStatus {
  running: boolean;
  jobs_scanned_today: number;
  applications_today: number;
  /** @nullable */
  last_run_at?: string | null;
}

export interface DashboardSummary {
  jobs_found_today: number;
  applications_total: number;
  interviews: number;
  offers: number;
  agent_running: boolean;
  onboarding_complete: boolean;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  job_found: 'job_found',
  applied: 'applied',
  interview: 'interview',
  offer: 'offer',
  rejected: 'rejected',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  company: string;
  timestamp: string;
  /** @nullable */
  match_score?: number | null;
}

export type ListJobsParams = {
status?: ListJobsStatus;
search?: string;
minMatch?: number;
};

export type ListJobsStatus = typeof ListJobsStatus[keyof typeof ListJobsStatus];


export const ListJobsStatus = {
  all: 'all',
  pending: 'pending',
  applied: 'applied',
  ignored: 'ignored',
} as const;

export type ListApplicationsParams = {
status?: ListApplicationsStatus;
};

export type ListApplicationsStatus = typeof ListApplicationsStatus[keyof typeof ListApplicationsStatus];


export const ListApplicationsStatus = {
  all: 'all',
  applied: 'applied',
  interview: 'interview',
  offer: 'offer',
  rejected: 'rejected',
} as const;