# lib/api-zod/src/generated/types/agentStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Job Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentStatus {
  running: boolean;
  jobs_scanned_today: number;
  applications_today: number;
  /** @nullable */
  last_run_at?: string | null;
}