# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Job Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  jobs_found_today: number;
  applications_total: number;
  interviews: number;
  offers: number;
  agent_running: boolean;
  onboarding_complete: boolean;
}