# lib/api-zod/src/generated/types/preferences.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Job Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { PreferencesWorkType } from './preferencesWorkType';

export interface Preferences {
  id: number;
  work_type: PreferencesWorkType;
  /** @nullable */
  city?: string | null;
  roles: string[];
  /** @nullable */
  min_salary?: number | null;
  updated_at?: string;
}