# lib/api-zod/src/generated/types/preferencesInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Job Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { PreferencesInputWorkType } from './preferencesInputWorkType';

export interface PreferencesInput {
  work_type: PreferencesInputWorkType;
  /** @nullable */
  city?: string | null;
  roles: string[];
  /** @nullable */
  min_salary?: number | null;
}