# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Job Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResult {
  jobs_found: number;
  message: string;
}